+++ src/test/java/domain/player/GamblerTest.java
@@ -2,6 +2,7 @@ package domain.player;
 
 import domain.betting.BettingAmount;
 import domain.card.Card;
+import java.math.BigDecimal;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,7 +14,7 @@ class GamblerTest {
     void 카드_정상적으로_받기_테스트() {
         // given
         Card card = new Card("2", "하트");
-        Gambler gambler = new Gambler("coco", new BettingAmount(20000));
+        Gambler gambler = new Gambler("coco", new BettingAmount(BigDecimal.valueOf(20000)));
 
         // when
         gambler.addCard(card);
@@ -30,7 +31,7 @@ class GamblerTest {
         Card card2 = new Card("K", "다이아몬드");
         Card card3 = new Card("K", "클로버");
 
-        Gambler gambler = new Gambler("coco", new BettingAmount(20000));
+        Gambler gambler = new Gambler("coco", new BettingAmount(BigDecimal.valueOf(20000)));
 
         // when
         gambler.addCard(card1);
@@ -49,7 +50,7 @@ class GamblerTest {
         Card card2 = new Card("K", "다이아몬드");
         Card card3 = new Card("K", "클로버");
 
-        Gambler gambler = new Gambler("coco", new BettingAmount(20000));
+        Gambler gambler = new Gambler("coco", new BettingAmount(BigDecimal.valueOf(20000)));
 
         // when
         gambler.addCard(card1);
